# apps/api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "LLM gateway API for the mobile app.",
        "title": "Daily Life Support API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "APIのヘルスチェック",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/llm/chat": {
            "post": {
                "description": "LLMにチャットリクエストを送信",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "LLM chat",
                "parameters": [
                    {
                        "description": "Chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llm.ChatInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/llm.ChatOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/llm/onboarding-plan": {
            "post": {
                "description": "オンボーディング入力値からAI提案を生成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding suggestion",
                "parameters": [
                    {
                        "description": "Onboarding plan request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/llm.ChatOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/llm/onboarding-plan-draft": {
            "delete": {
                "description": "保存済みDraftを削除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding draft delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "オンボーディング入力値から構造化プランを生成しDraft保存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding draft create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Onboarding draft request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanDraftOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/llm/onboarding-plan-draft/apply": {
            "post": {
                "description": "保存済みDraftを実DBに反映",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding draft apply",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanApplyOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/llm/onboarding-plan-apply": {
            "post": {
                "description": "オンボーディング入力値からAIプランを生成してDBに保存",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding plan generate and apply",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Onboarding plan apply request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanApplyOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/llm/onboarding-plan-structured": {
            "post": {
                "description": "オンボーディング入力値からDB投入用の構造化プランJSONを生成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llm"
                ],
                "summary": "Onboarding structured plan",
                "parameters": [
                    {
                        "description": "Onboarding structured plan request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/llm.OnboardingPlanInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/httphandler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httphandler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "llm.ChatInput": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/llm.Message"
                    }
                },
                "model": {
                    "type": "string"
                },
                "temperature": {
                    "type": "number"
                }
            }
        },
        "llm.ChatOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                }
            }
        },
        "llm.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "llm.OnboardingPlanInput": {
            "type": "object",
            "properties": {
                "heightCm": {
                    "type": "number"
                },
                "occupation": {
                    "type": "string"
                },
                "targetDurationUnit": {
                    "description": "week or month",
                    "type": "string"
                },
                "targetDurationValue": {
                    "type": "integer"
                },
                "targetWeightKg": {
                    "type": "number"
                },
                "trainingMode": {
                    "description": "GYM or BODYWEIGHT",
                    "type": "string"
                },
                "weightKg": {
                    "type": "number"
                }
            }
        },
        "llm.OnboardingPlanDraftOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "plan": {
                    "type": "object"
                }
            }
        },
        "llm.OnboardingPlanApplyOutput": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}